3 services done

+++ _BLL/IUserManager.cs
@@ -9,8 +9,8 @@ namespace BLL
     public interface IUserManager
     {
         List<UserAccount> GetUsers();
-        float GetAvailableAmount(int id);
-        int ReloadMoneyUID(int id, double amount);
+        int GetAvailableAmount(int id);
+        int ReloadMoneyUID(int id, int amount);
         
     }
 }
