+++ _DAL/IUserDB.cs
@@ -8,7 +8,7 @@ namespace DAL
     public interface IUserDB
     {
         List<UserAccount> GetUsers();
-        float GetAvailableAmount(int id);
-        int ReloadMoneyUID(int id, double amount);
+        int GetAvailableAmount(int id);
+        int ReloadMoneyUID(int id, int amount);
     }
 }
